requirements.txt
django
djangorestframework
django-cors-headers
requests
pyjwt 
djangorestframework-simplejwt